requirements.txt
streamlit>=1.22.0
sentence-transformers>=2.2.2
numpy>=1.22.0
scikit-learn>=1.1.0
torch>=1.12.0
requests>=2.28.1
python-dotenv>=0.19.0
flask==2.3.3
flask-cors==4.0.0
chromadb>=0.4.22
langchain>=0.1.0